+++ core/graphicsManager/graphicsLinker.cpp
@@ -38,7 +38,7 @@ void GraphicsLinker::createRenderPass(VkDevice device) {
         moon::utils::Attachments::imageDescription(imageInfo.Format, VK_IMAGE_LAYOUT_PRESENT_SRC_KHR)
     };
 
-    utils::vkDefault::SubpassInfos subpassInfos = utils::vkDefault::subpassInfos(attachments.size());
+    utils::vkDefault::SubpassInfos subpassInfos = utils::vkDefault::subpassInfos(static_cast<uint32_t>(attachments.size()));
 
     utils::vkDefault::RenderPass::SubpassDependencies dependencies;
     dependencies.push_back(VkSubpassDependency{});
